hide functionality if not logged in

+++ src/App.js
@@ -34,11 +34,13 @@ function App() {
   const URL = "https://project-3-backend-hard.herokuapp.com/all/" // Back-end Heroku link
 
   const getRestaurant = async () => {
+    if(user) return
     const response = await fetch(URL)
     const data = await response.json();
     setRestaurant(data);
   }
   const createRestaurant = async (restaurant) => {
+    if(!user) return
     await fetch(URL, {
       method: "POST",
       headers: {
@@ -50,6 +52,7 @@ function App() {
   };
 
   const updateRestaurant = async (updatedRestaurant, restaurant) => {
+    if(!user) return
     await fetch(URL + restaurant, {
       method: "PUT",
       headers: {
@@ -61,6 +64,7 @@ function App() {
   };
   
   const deleteRestaurant = async (restaurant) => {
+    if(!user) return
     await fetch(URL + restaurant, { method: "DELETE" });
     getRestaurant();
   };
@@ -106,7 +110,10 @@ function App() {
         <Route
           path="/restaurant/:id"
           render={(renderProps) => (
-            <Show restaurant={restaurant} {...renderProps}
+            <Show 
+            user={user}
+            restaurant={restaurant} 
+            {...renderProps}
             deleteRestaurant={deleteRestaurant}
             updateRestaurant={updateRestaurant}
             />
